Authentication pages html and css changes for displaying errors (#25)
* Reformated error html code. Added css attributes to error messages to give a better stlye. Fixed errors causing register and login form content from spilling out of forms. Lastly changed some of the error messages so they werent so long.

* Error #7 fix and html cleanup

Fixed the Swift_Transport error the mail needed to be setup. But everyone needs to add an email, password, and port # to the .env file or get the email info from me for the email I just created. Made changes to the reset html pages to have the error messages contain the same css style. ------->Or change the .env file's (MAIL_Driver=log) then the reset password page will log the email instead of trying to email it. The log is found in config/logs/laravel.log at the very bottom it will have the reset link.<-------