Added removing member and edited file tab
Made it so admins can kick non admin users from the project. The file tab now includes create file and the search bar. Tweaked some of the css so the data inside of tabs are staying within their tabs. Also moved some of the routes to a project members controller. So all things located on the public project page are located in the project controller and everything located on the project private members page is located on the new controller.